feat(community): refactor search input and community card components
