+++ composables/useCommunities.ts
@@ -1,4 +1,4 @@
-import { communities } from "./../public/data/communities";
+import { communities } from "public/data/communities";
 export const useCommunities = () => {
   return useState("communities", () => communities);
 };
